requirements.txt
djangorestframework, django-cors-headers, pytest, djangorestframework-simplejwt
